implementation

+++ common/ble_imu/ble_imu_types.h
@@ -53,7 +53,7 @@ typedef struct
             uint8_t rate;
             uint8_t flags;
             uint8_t stream_enable;
-            uint8_t sampling_time;
+            uint16_t sampling_time;
         } record;
     };
 } __attribute__((packed)) command_payload_t;
